move digest to separate file
